requirements.txt
numpy==1.21.0
pyrtlsdr==0.2.1
scipy==1.7.0
sounddevice==0.4.1
